# site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Run402 API",
    "version": "2.0.0",
    "description": "Full-stack infrastructure for AI agents. Postgres, REST API, auth, file storage, static site hosting, serverless functions. Pay-per-tier via x402 USDC on Base — subscribe once, then use wallet signatures for all operations within your tier.",
    "contact": {
      "name": "Run402",
      "url": "https://run402.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.run402.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/projects/v1": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "operationId": "listProjects",
        "description": "List projects. Wallet auth sees own projects; admin auth (ADMIN_KEY or admin wallet SIWx) sees all projects. Supports cursor-based pagination.",
        "security": [
          {
            "walletAuth": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WalletHeader"
          },
          {
            "$ref": "#/components/parameters/SignatureHeader"
          },
          {
            "$ref": "#/components/parameters/TimestampHeader"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            },
            "description": "Maximum number of projects to return (default 50, max 200)"
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination (from previous response's next_cursor)"
          }
        ],
        "responses": {
          "200": {
            "description": "Project list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectCreated"
                      }
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Provision a new project",
        "operationId": "createProject",
        "description": "Creates a new database project. Requires wallet auth (EIP-4361 signature). Free within your active tier. The wallet must have an active tier subscription.",
        "security": [
          {
            "walletAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WalletHeader"
          },
          {
            "$ref": "#/components/parameters/SignatureHeader"
          },
          {
            "$ref": "#/components/parameters/TimestampHeader"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Optional project name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreated"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/projects/v1/{id}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Archive (soft-delete) a project",
        "operationId": "deleteProject",
        "description": "Archives and soft-deletes a project. Accepts service_key or admin auth (ADMIN_KEY or admin wallet SIWx).",
        "security": [
          {
            "serviceKey": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project purged (terminal delete)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "purged"
                    },
                    "project_id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/subdomains/v1": {
      "post": {
        "tags": [
          "Subdomains"
        ],
        "summary": "Claim or reassign a subdomain",
        "operationId": "createSubdomain",
        "description": "Maps a custom subdomain ({name}.run402.com) to a deployment.",
        "security": [
          {
            "serviceKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "deployment_id"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Subdomain name (e.g. 'myapp' for myapp.run402.com)"
                  },
                  "deployment_id": {
                    "type": "string",
                    "description": "Deployment ID to point to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subdomain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subdomain"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Subdomains"
        ],
        "summary": "List subdomains",
        "operationId": "listSubdomains",
        "description": "List subdomains. service_key sees the project's subdomains; admin auth (ADMIN_KEY or admin wallet SIWx) sees all subdomains.",
        "security": [
          {
            "serviceKey": []
          },
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subdomain list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subdomains": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Subdomain"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/subdomains/v1/{name}": {
      "get": {
        "tags": [
          "Subdomains"
        ],
        "summary": "Look up a subdomain",
        "operationId": "getSubdomain",
        "description": "Public lookup — no auth required.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subdomain details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subdomain"
                }
              }
            }
          },
          "404": {
            "description": "Subdomain not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Subdomains"
        ],
        "summary": "Release a subdomain",
        "operationId": "deleteSubdomain",
        "description": "Releases a subdomain. Accepts service_key or admin auth (ADMIN_KEY or admin wallet SIWx).",
        "security": [
          {
            "serviceKey": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subdomain released",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "deleted"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Project anon_key (JWT). Returned when you provision a project."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access_token from /auth/v1/token. Pass as Authorization: Bearer <token>."
      },
      "serviceKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Project service_key (JWT with service_role). Returned when you provision a project. Pass as Authorization: Bearer <service_key>."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Server admin key. Required for admin operations."
      },
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "x402 payment header. The client library (e.g. @anthropic-ai/sdk x402 facilitator) handles this automatically. Payment is in USDC on Base (mainnet eip155:8453 or testnet eip155:84532)."
      },
      "walletAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Run402-Wallet",
        "description": "EIP-4361 wallet authentication. Requires three headers: X-Run402-Wallet (wallet address), X-Run402-Signature (signed message), X-Run402-Timestamp (unix timestamp). The wallet must have an active tier subscription."
      }
    },
    "parameters": {
      "WalletHeader": {
        "name": "X-Run402-Wallet",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Ethereum wallet address (0x...)"
      },
      "SignatureHeader": {
        "name": "X-Run402-Signature",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "EIP-4361 signature of the message 'run402:{method}:{path}:{timestamp}'"
      },
      "TimestampHeader": {
        "name": "X-Run402-Timestamp",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Unix timestamp (must be within 30 seconds of server time)"
      }
    },
    "schemas": {
...
      "ProjectCreated": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "anon_key": {
            "type": "string",
            "description": "JWT for client-side access (apikey header)"
          },
          "service_key": {
            "type": "string",
            "description": "JWT for admin access (Authorization: Bearer)"
          },
          "schema_slot": {
            "type": "string"
          },
          "tier": {
            "type": "string",
            "enum": [
              "prototype",
              "hobby",
              "team"
            ]
          },
          "lease_expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Subdomain": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "deployment_id": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "deployment_url": {
            "type": "string",
            "format": "uri"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication. For wallet-auth endpoints, provide X-Run402-Wallet, X-Run402-Signature, and X-Run402-Timestamp headers. For project endpoints, provide apikey header (anon_key) or Authorization: Bearer (service_key/access_token).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}